Start user repo tests

+++ src/main/java/com/clouway/bankapp/adapter/datastore/DatastoreTranslationRepository.kt
@@ -8,11 +8,11 @@ class DatastoreTranslationRepository : TransactionRepository {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getUserTransactions(id: Int, page: Int, pageSize: Int): List<Transaction> {
+    override fun getUserTransactions(id: Long, page: Int, pageSize: Int): List<Transaction> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getUserTransactions(id: Int): List<Transaction> {
+    override fun getUserTransactions(id: Long): List<Transaction> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }
