+++ src/main/java/com/clouway/bankapp/core/TransactionRepository.kt
@@ -4,7 +4,7 @@ interface TransactionRepository {
 
     fun save(transaction: Transaction)
 
-    fun getUserTransactions(id: Int, page: Int, pageSize: Int): List<Transaction>
-    fun getUserTransactions(id: Int): List<Transaction>
+    fun getUserTransactions(id: Long, page: Int, pageSize: Int): List<Transaction>
+    fun getUserTransactions(id: Long): List<Transaction>
 
 }
