+++ src/main/java/com/clouway/bankapp/core/User.kt
@@ -1,3 +1,3 @@
 package com.clouway.bankapp.core
 
-data class User (val id: Int, val username: String, val password: String)
+data class User (val id: Long, val username: String, val password: String)
